# docs/swagger.yaml
basePath: /
definitions:
  models.Delivery:
    properties:
      address:
        type: string
      city:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      region:
        type: string
      zip:
        type: string
    required:
    - address
    - city
    - email
    - name
    - phone
    - region
    - zip
    type: object
  models.Item:
    properties:
      brand:
        type: string
      chrt_id:
        type: integer
      name:
        type: string
      nm_id:
        type: integer
      price:
        type: integer
      rid:
        type: string
      sale:
        maximum: 100
        minimum: 0
        type: integer
      size:
        type: integer
      status:
        type: integer
      total_price:
        type: integer
      track_number:
        type: string
    required:
    - brand
    - chrt_id
    - name
    - nm_id
    - price
    - rid
    - sale
    - size
    - status
    - total_price
    - track_number
    type: object
  models.Order:
    properties:
      customer_id:
        type: string
      date_created:
        type: string
      delivery:
        $ref: '#/definitions/models.Delivery'
      delivery_service:
        type: string
      entry:
        type: string
      internal_signature:
        type: string
      items:
        items:
          $ref: '#/definitions/models.Item'
        minItems: 1
        type: array
      locale:
        type: string
      oof_shard:
        type: string
      order_uid:
        type: string
      payment:
        $ref: '#/definitions/models.Payment'
      shardkey:
        type: string
      sm_id:
        type: integer
      track_number:
        type: string
    required:
    - customer_id
    - date_created
    - delivery
    - delivery_service
    - entry
    - items
    - locale
    - oof_shard
    - order_uid
    - payment
    - shardkey
    - sm_id
    - track_number
    type: object
  models.Payment:
    properties:
      amount:
        type: integer
      bank:
        type: string
      currency:
        type: string
      custom_fee:
        minimum: 0
        type: integer
      delivery_cost:
        minimum: 0
        type: integer
      goods_total:
        type: integer
      payment_dt:
        type: integer
      provider:
        type: string
      request_id:
        type: string
      transaction:
        type: string
    required:
    - amount
    - bank
    - currency
    - delivery_cost
    - goods_total
    - payment_dt
    - provider
    - transaction
    type: object
host: localhost:8081
info:
  contact: {}
  description: This is a order microservice
  title: order-manager
  version: "1.0"
paths:
  /order/{order_uid}:
    get:
      parameters:
      - description: Order UID
        in: path
        name: order_uid
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Bad request
          schema: {}
        "404":
          description: Not found
          schema: {}
      summary: Get order by UID
swagger: "2.0"